requirements.txt
Django==1.6.3
django-pagedown==0.1.0
djangorestframework==2.3.13
wsgiref==0.1.2
